requirements.txt
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.1
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
black==26.1.0
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.5
click==8.3.1
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distlib==0.4.0
django==6.0.3
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.0
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.17
idna==3.11
ipykernel==7.2.0
ipython==9.11.0
ipython-pygments-lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-client==8.8.0
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.4
jupyterlab==4.5.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
lark==1.3.1
lxml==6.0.2
markupsafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
mypy-extensions==1.1.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.10.0
notebook==7.5.4
notebook-shim==0.2.4
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pandocfilters==1.5.1
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
pre-commit==4.5.1
prometheus-client==0.24.1
prompt-toolkit==3.0.52
psutil==7.2.2
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==3.0
pygments==2.19.2
pyjwt==2.11.0
python-dateutil==2.9.0.post0
python-discovery==1.1.0
python-json-logger==4.0.0
pytokens==0.4.1
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
ruff==0.15.4
send2trash==2.1.0
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typing-extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
virtualenv==21.1.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
